esp_claim: Add support for "Assisted Claiming"
This is specifically for simplifying ESP32 claiming workflow,
wherein the phone apps will assist the node for claiming during
Wi-Fi provisioning. There would be no need of using CLI for claiming.

This will work only for BLE provisioning, as internet connectivity
will be required for claiming, and it may not be possible with
SoftAP.