bugfix: Callback for persistent parameters wasn't getting invoked
The RainMaker core has special handling for the "name" type of parameter
and so, it does not call the application registered callback for this
parameter. However, the internal check was erroneous and the callback
was getting skipped even when the param type was NULL. The check has
now been fixed.