wifi_prov: Fixed a bug which was causing name/pop to change on every reboot
The provisioning name (SSID/BLE device name) and PoP are generated from
the general purpose random bytes stored in the factory NVS partition.
If a node hasn't completed claiming, these random bytes was getting re-generated
on a reboot, thereby causing the name and pop to change, which is undesirable.

With this fix, the random bytes will be generated only if not done already.

Note that an erase_flash will erase these bytes as well and so,
the provisioning name and PoP will still change in that case.