esp_rmaker_client_data: Allow to use MQTT Host from config even when claiming is disabled
Now that we allow reading private key and certificate from secure cert
partition too, the node may not have MQTT host in nvs even in production
cases wherein claiming is disabled. This was causing errors in MQTT
init. This new config option forces firmware to use it from NVS.

This also helps in cases where the same node is to be connected to a
different server after an OTA.