esp_rmaker_param: Add support for Simple Time Series Data
This is a simpler variant of time series data, which allows fetching
only raw values and not any aggregates like average, min, max, etc.
This option is cost efficient and is recommended if the application
requires only raw values.

Use PROP_FLAG_SIMPLE_TIME_SERIES instead of PROP_FLAG_TIME_SERIES in the
parameter properties flags to enable this.