esp_rmaker_claim: Fetch MQTT Host from assisted claiming service
For assisted claiming, the node itself does not connect to the
claiming service directly. Instead, it just creates and parses payloads which
are sent/received by phone apps on its behalf. Since the phone app
can get certificates from any claiming service, it will be better
to have the MQTT Host URL also part of the response, so that there
is no mismatch between certificate and MQTT Host URL.