esp_rainmaker/core: fix the memory mgmt issues in client data retrieval APIs
esp_rmaker_get_client_cert() and esp_rmaker_get_client_key() were
returning different types of memory. If secure cert config is enabled
then it would returned the memory mapped address, if not enabled or if
this secure cert based retrieval fails then it returns the address on
heap so it is pretty difficult to figure out whether to free or not.

With this change if secure cert option is enabled then it allocates the
buffer and returns that allocated buffers, making APIs to always return
the address on heap so that it can be freed with free().

Also added documentation on how to free the returned buffers.

Also, fix mem leaks and perform null check on dynamic allocation.