examples/matter/matter_light: set default values using set_val api
Variable names differ between main and release/v1.4 branches for setting
the member attributes in feature struct, so use the attribute::set_val()
api as a workaround for compatibility.

esp-matter!1002 added the feature config and we updated the rainmaker
examples, but then there were more changes went in which broke the
APIs/structs and was fixed in esp-matter!1146 which again broke the
rainmaker examples.

This fixes the current esp-matter:main and esp-rainmaker:master builds
for rainmaker+matter examples.