rmaker_app_network: Better handling of provisioning failures
Remove the retry logic for Wi-Fi disconnected events while in
provisioning mode because it is handled internally by the
network_provisioning component. Now, depending on the Kconfig options,
the firmeare will try Wi-Fi connection one or more time and then stop.
Phone apps can reset the state and retry provisioning on the same
session.

Also, reduced the default connection retry count from 5 to 3 so that
users would have to wait for lesser time for retrying